Refactor DCAN code
